+++ src/components/AppBarButton.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled, { withTheme } from 'styled-components'
 import Icon from './Icon'
-import theme, { ThemeInterface } from '../theme/theme'
-import icons from '../theme/icons'
+import theme, { ThemeInterface } from 'theme/theme'
+import icons from 'theme/icons'
 
 interface AppBarButtonProps {
   title: string
