+++ src/components/ResultsPanel.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import RepoList from './RepoList'
-import { StarProvider } from '../state/star-context'
+import { StarProvider } from 'state/star-context'
 
 const ResultsPanel = () => {
   return (
