+++ src/components/UserAuthenticator.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
-import logo from '../assets/vectors/logo.svg'
+import logo from 'assets/vectors/logo.svg'
 import Button from './Button'
 
 enum AuthState {
