+++ src/components/__tests__/RepoList.test.tsx
@@ -4,11 +4,11 @@ import RepoList, {
   sanitizeString,
   stringToArray
 } from '../RepoList'
-import { render as renderrtl, cleanup } from '../utils/test-utils'
+import { render as renderrtl, cleanup } from 'utils/test-utils'
 import 'jest-dom/extend-expect'
-import { StarProvider } from '../../state/star-context'
-import { StarredRepo } from '../../types/GithubTypes'
-import { SearchProvider } from '../../state/search-context'
+import { StarProvider } from 'state/star-context'
+import { StarredRepo } from 'types/GithubTypes'
+import { SearchProvider } from 'state/search-context'
 
 afterEach(() => cleanup())
