+++ src/components/__tests__/UserAuthenticator.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import UserAuthenticator from '../UserAuthenticator'
-import { render, cleanup, fireEvent } from '../utils/test-utils'
-import userData from '../../mock-data/user.json'
+import { render, cleanup, fireEvent } from 'utils/test-utils'
+import userData from 'mock-data/user.json'
 
 const authorizeMock = jest.fn()
